requirements.txt
openai>=1.12.0
open_spiel>=1.5
numpy>=1.23
scipy>=1.10
restrictedpython>=6.2
cvxpy>=1.3
ecos>=2.0
osqp>=0.6
tqdm>=4.66
python-dotenv